🚚 refactor: move skeleton directory under header
