+++ src/components/header/UserInfo.tsx
@@ -2,11 +2,11 @@
 
 import type { User } from '@/types/user';
 import Image from 'next/image';
-import useWindowSize from '@/app/(with-header-sidebar)/mydashboard/_hooks/useWindowSize';
 import { useState, useEffect } from 'react';
-import UserInfoSkeleton from '../skeleton/UserInfoSkeleton';
 import styles from './UserInfo.module.css';
+import UserInfoSkeleton from './skeleton/UserInfoSkeleton';
 
+// TODO 로그인 로직 완료 후 user정보 가져오기
 const user: User = {
   id: 1,
   email: 'heejin@gmail.com',
@@ -19,7 +19,6 @@ const user: User = {
 
 export default function UserInfo() {
   const { email, nickname, profileImageUrl } = user;
-  const { isMobile } = useWindowSize();
 
   const [colors, setColors] = useState<{
     randomColor: string;
